fixed data types

+++ dbt_subprojects/dex/models/trades/ronin/platforms/tamadotmeme_v1_ronin_base_trades.sql
@@ -116,11 +116,11 @@ select
 , cast (token_bought_symbol as varchar) as token_bought_symbol
 , cast (token_sold_symbol as varchar) as token_sold_symbol
 , cast (token_pair as varchar) as token_pair
-, cast (token_bought_amount as uint256) as token_bought_amount
-, cast (token_sold_amount as uint256) as token_sold_amount
-, cast (token_bought_amount_raw as double) as token_bought_amount_raw
-, cast (token_sold_amount_raw as double) as token_sold_amount_raw
-, cast (amount_usd as uint256) as amount_usd
+, cast (token_bought_amount as double) as token_bought_amount
+, cast (token_sold_amount as double) as token_sold_amount
+, cast (token_bought_amount_raw as uint256) as token_bought_amount_raw
+, cast (token_sold_amount_raw as uint256) as token_sold_amount_raw
+, cast (amount_usd as double) as amount_usd
 , cast (token_bought_address as varbinary) as token_bought_address
 , cast (token_sold_address as varbinary) as token_sold_address
 , cast (taker as varbinary) as taker
